refactor(si): satisfy the repo's lint, size and complexity gates
CI's oxlint --type-aware stage reported 25 errors that never appeared
locally, because the local ci:check stopped earlier on the Windows CRLF
false alarm. Fixed at the root rather than by relaxing any rule.

Type assertions removed by making the types true instead of asserting
them: drizzle's $inferSelect replaces hand-written row interfaces, the
projection helpers take typed parameters instead of unknown, readJson uses
a type predicate so the check and the narrowing are one statement, and the
counter snapshot is written as an explicit literal — adding a counter to
the union without adding it there is now a build error rather than a
silent zero.

Three modules exceeded the 400-line limit and dispatch had a complexity of
62 against a ceiling of 40 with six parameters against five. Split along
real seams rather than arbitrary cut points: ledger-store (accounting) and
cost (the read side of the ledger) leave store and service; projections
(what we expose), http (response shaping) and router (the route table)
leave api, which keeps only the posture — feature gate, logging, and
turning any failure into a sanitized response. dispatch now takes one
context object and splits into reads and operations, so "what can change
state" is a readable list instead of a branch buried in a 60-way
conditional.

Also: toSorted over sort, Array.from over a useless spread, and the last
unused exports removed.

Local gate now matches CI: prettier, knip, tsc, oxlint --type-aware all
clean, 878 tests pass.
